+++ src/read/stream.rs
@@ -6,9 +6,7 @@ use crate::read::{
     ZipFile, ZipFileData, ZipResult, central_header_to_zip_file_inner, make_symlink,
 };
 use crate::result::ZipError;
-use crate::spec::Magic;
-use crate::spec::Pod;
-use crate::spec::{FixedSizeBlock, ZipCentralEntryBlock, ZipLocalEntryBlock};
+use crate::spec::{FixedSizeBlock, Magic, Pod, ZipCentralEntryBlock, ZipLocalEntryBlock, is_dir};
 use indexmap::IndexMap;
 use std::borrow::Cow;
 use std::io::{self, Read};
@@ -191,10 +189,7 @@ impl ZipStreamFileMetadata {
 
     /// Returns whether the file is actually a directory
     pub fn is_dir(&self) -> bool {
-        self.name()
-            .chars()
-            .next_back()
-            .is_some_and(|c| c == '/' || c == '\\')
+        is_dir(self.name())
     }
 
     /// Returns whether the file is a regular file
@@ -235,14 +230,17 @@ pub fn read_zipfile_from_stream<R: Read>(reader: &mut R) -> ZipResult<Option<Zip
     // finished iterating over all the actual files).
     /* TODO: smallvec? */
 
-    let mut block = ZipLocalEntryBlock::zeroed();
-    reader.read_exact(block.as_bytes_mut())?;
+    let mut magic_buf = [0; size_of::<u32>()];
+    reader.read_exact(&mut magic_buf)?;
 
-    match block.magic().from_le() {
+    match Magic::from_le_bytes(magic_buf) {
         Magic::LOCAL_FILE_HEADER_SIGNATURE => (),
         Magic::CENTRAL_DIRECTORY_HEADER_SIGNATURE => return Ok(None),
         _ => return Err(ZipLocalEntryBlock::WRONG_MAGIC_ERROR),
     }
+
+    let mut block = ZipLocalEntryBlock::zeroed();
+    reader.read_exact(block.as_bytes_mut())?;
 
     let block = block.from_le();
 
@@ -283,14 +281,17 @@ pub fn read_zipfile_from_stream_with_compressed_size<R: io::Read>(
     reader: &mut R,
     compressed_size: u64,
 ) -> ZipResult<Option<ZipFile<'_, R>>> {
-    let mut block = ZipLocalEntryBlock::zeroed();
-    reader.read_exact(block.as_bytes_mut())?;
+    let mut magic_buf = [0; size_of::<u32>()];
+    reader.read_exact(&mut magic_buf)?;
 
-    match block.magic().from_le() {
+    match Magic::from_le_bytes(magic_buf) {
         Magic::LOCAL_FILE_HEADER_SIGNATURE => (),
         Magic::CENTRAL_DIRECTORY_HEADER_SIGNATURE => return Ok(None),
         _ => return Err(ZipLocalEntryBlock::WRONG_MAGIC_ERROR),
     }
+
+    let mut block = ZipLocalEntryBlock::zeroed();
+    reader.read_exact(block.as_bytes_mut())?;
 
     let block = block.from_le();